+++ blueprints/codes/eurocode/national_annex/nl/nen_en_1992_1_1_a1_2020/chapter_4_durability_and_cover/formula_4_1.py
@@ -1,6 +1,6 @@
 """Formula 4.1 from NEN-EN 1992-1-1:2005+A1:2015+NB:2016+A1:2020: Chapter 4 - Durability and cover to reinforcement."""
 
-from blueprints.codes.eurocode.nen_en_1992_1_1_a1_2020 import NEN_EN_1992_1_1_A1_2020
+from blueprints.codes.eurocode.national_annex.nl.nen_en_1992_1_1_a1_2020 import NEN_EN_1992_1_1_A1_2020
 from blueprints.codes.formula import Formula
 from blueprints.codes.latex_formula import LatexFormula
 from blueprints.type_alias import MM
