+++ tests/codes/eurocode/national_annex/nl/nen_en_1992_1_1_a1_2020/chapter_4_durability_and_cover/test_formula_4_2.py
@@ -2,7 +2,9 @@
 
 import pytest
 
-from blueprints.codes.eurocode.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.formula_4_2 import Form4Dot2MinimumConcreteCover
+from blueprints.codes.eurocode.national_annex.nl.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.formula_4_2 import (
+    Form4Dot2MinimumConcreteCover,
+)
 from blueprints.validations import NegativeValueError
