+++ tests/codes/eurocode/national_annex/nl/nen_en_1992_1_1_a1_2020/chapter_4_durability_and_cover/test_table_4_4n.py
@@ -1,8 +1,8 @@
-"""Testing formula 4.5N of NEN-EN 1992-1-1:2005+A1:2015+NB:2016+A1:2020."""
+"""Testing formula 4.4N of NEN-EN 1992-1-1:2005+A1:2015+NB:2016+A1:2020."""
 
 import pytest
 
-from blueprints.codes.eurocode.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.table_4_1 import (
+from blueprints.codes.eurocode.national_annex.nl.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.table_4_1 import (
     Carbonation,
     Chemical,
     Chloride,
@@ -10,19 +10,19 @@ from blueprints.codes.eurocode.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_
     FreezeThaw,
     Table4Dot1ExposureClasses,
 )
-from blueprints.codes.eurocode.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.table_4_3 import (
+from blueprints.codes.eurocode.national_annex.nl.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.table_4_3 import (
     ConcreteMaterial,
     ConcreteStrengthClass,
     Table4Dot3ConcreteStructuralClass,
 )
-from blueprints.codes.eurocode.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.table_4_5n import (
-    Table4Dot5nMinimumCoverDurabilityPrestressingSteel,
+from blueprints.codes.eurocode.national_annex.nl.nen_en_1992_1_1_a1_2020.chapter_4_durability_and_cover.table_4_4n import (
+    Table4Dot4nMinimumCoverDurabilityReinforcementSteel,
 )
 from blueprints.type_alias import MM
 
 
-class TestTable4Dot5nMinimumCoverDurabilityPrestressingSteel:
-    """Validation for table 4.5N from NEN-EN 1992-1-1:2005+A1:2015+NB:2016+A1:2020."""
+class TestTable4Dot4nMinimumCoverDurabilityReinforcementSteel:
+    """Validation for table 4.4 from NEN-EN 1992-1-1:2005+A1:2015+NB:2016+A1:2020."""
 
     @pytest.mark.parametrize(
         ("exposure_classes", "structural_class", "expected_result"),
@@ -30,17 +30,17 @@ class TestTable4Dot5nMinimumCoverDurabilityPrestressingSteel:
             (
                 Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.XD1, ChlorideSeawater.XS3, FreezeThaw.XF4, Chemical.XA1),
                 6,
-                55,
+                50,
             ),
             (
                 Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.XD3, ChlorideSeawater.XS2, FreezeThaw.XF4, Chemical.XA1),
                 5,
-                50,
+                45,
             ),
             (
                 Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.XD1, ChlorideSeawater.XS2, FreezeThaw.XF4, Chemical.XA1),
                 5,
-                50,
+                45,
             ),
             (
                 Table4Dot1ExposureClasses(Carbonation.NA, Chloride.NA, ChlorideSeawater.NA, FreezeThaw.NA, Chemical.NA),
@@ -50,22 +50,22 @@ class TestTable4Dot5nMinimumCoverDurabilityPrestressingSteel:
             (
                 Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.NA, ChlorideSeawater.NA, FreezeThaw.XF4, Chemical.XA1),
                 3,
-                25,
+                20,
             ),
             (
                 Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.XD1, ChlorideSeawater.NA, FreezeThaw.XF4, Chemical.XA1),
                 2,
-                30,
+                25,
             ),
             (
                 Table4Dot1ExposureClasses(Carbonation.XC4, Chloride.NA, ChlorideSeawater.NA, FreezeThaw.XF4, Chemical.XA1),
                 1,
-                20,
+                15,
             ),
             (
-                Table4Dot1ExposureClasses(Carbonation.XC1, Chloride.NA, ChlorideSeawater.NA, FreezeThaw.XF3, Chemical.XA1),
+                Table4Dot1ExposureClasses(Carbonation.XC1, Chloride.NA, ChlorideSeawater.NA, FreezeThaw.XF2, Chemical.XA1),
                 5,
-                25,
+                20,
             ),
         ],
     )
@@ -73,19 +73,19 @@ class TestTable4Dot5nMinimumCoverDurabilityPrestressingSteel:
         self, exposure_classes: Table4Dot1ExposureClasses, structural_class: Table4Dot3ConcreteStructuralClass, expected_result: MM
     ) -> None:
         """Test the evaluation of the result."""
-        table_4_5 = Table4Dot5nMinimumCoverDurabilityPrestressingSteel(
+        table_4_4 = Table4Dot4nMinimumCoverDurabilityReinforcementSteel(
             exposure_classes=exposure_classes,
             structural_class=structural_class,
         )
 
-        assert table_4_5 == pytest.approx(expected=expected_result, rel=1e-4)
+        assert table_4_4 == pytest.approx(expected=expected_result, rel=1e-4)
 
     def test_evaluation_valid_structural_class_type(self) -> None:
         """Test if the evaluation raises TypeError when the structural class is not an integer."""
         exposure_classes = Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.XD1, ChlorideSeawater.XS3, FreezeThaw.XF4, Chemical.XA1)
         structural_class = Table4Dot3ConcreteStructuralClass(exposure_classes, 50, ConcreteMaterial(ConcreteStrengthClass("C20/25")), True, False)
 
-        assert Table4Dot5nMinimumCoverDurabilityPrestressingSteel(exposure_classes, structural_class) == 40
+        assert Table4Dot4nMinimumCoverDurabilityReinforcementSteel(exposure_classes, structural_class) == 35
 
     def test_evaluation_invalid_structural_class_type(self) -> None:
         """Test if the evaluation raises TypeError when the structural class is not an integer."""
@@ -93,13 +93,21 @@ class TestTable4Dot5nMinimumCoverDurabilityPrestressingSteel:
         structural_class = "S4"
 
         with pytest.raises(TypeError):
-            Table4Dot5nMinimumCoverDurabilityPrestressingSteel(exposure_classes, structural_class)  # type: ignore[arg-type]
+            Table4Dot4nMinimumCoverDurabilityReinforcementSteel(exposure_classes, structural_class)  # type: ignore[arg-type]
+
+    def test_evaluation_invalid_structural_class_value(self) -> None:
+        """Test if the evaluation raises ValueError when the structural class is not between 1 and 6."""
+        exposure_classes = Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.XD1, ChlorideSeawater.XS3, FreezeThaw.XF4, Chemical.XA1)
+        structural_class = 7
+
+        with pytest.raises(ValueError):
+            Table4Dot4nMinimumCoverDurabilityReinforcementSteel(exposure_classes, structural_class)  # type: ignore[arg-type]
 
     @pytest.mark.parametrize(
         ("representation", "expected_result"),
         [
-            ("complete", r"c_{min,dur} = structural class S3 & exposure classes (XC2, XF4, XA1) = 25"),
-            ("short", "c_{min,dur} = 25"),
+            ("complete", r"c_{min,dur} = structural class S3 \& exposure classes (XC2, XF4, XA1) = 20"),
+            ("short", "c_{min,dur} = 20"),
         ],
     )
     def test_latex(self, representation: str, expected_result: str) -> None:
@@ -108,7 +116,7 @@ class TestTable4Dot5nMinimumCoverDurabilityPrestressingSteel:
         exposure_classes = Table4Dot1ExposureClasses(Carbonation.XC2, Chloride.NA, ChlorideSeawater.NA, FreezeThaw.XF4, Chemical.XA1)
         structural_class = Table4Dot3ConcreteStructuralClass(exposure_classes, 50, ConcreteMaterial(ConcreteStrengthClass("C20/25")), True, False)
 
-        c_min_dur = Table4Dot5nMinimumCoverDurabilityPrestressingSteel(
+        c_min_dur = Table4Dot4nMinimumCoverDurabilityReinforcementSteel(
             exposure_classes=exposure_classes,
             structural_class=structural_class,
         ).latex()
