+++ tests/codes/eurocode/national_annex/nl/nen_en_1992_1_1_a1_2020/chapter_8_detailing_of_reinforcement_and_prestressing_tendons/__init__.py
